fix: compiler warning

+++ bluetoothserver/utils.cpp
@@ -81,7 +81,7 @@ std::RuntimeExceptionWithBacktrace::RuntimeExceptionWithBacktrace(const std::str
 	char ** messages = backtrace_symbols(array, size);    
 
 	// skip first stack frame (points here)
-	for (int i = 1; i < size && messages != NULL; ++i) {
+	for (size_t i = 1; i < size && messages != NULL; ++i) {
 		char *mangled_name = 0, *offset_begin = 0, *offset_end = 0;
 
 		// find parantheses and +address offset surrounding mangled name
